refactor(private-world): relocate reducer and service
